Reorganise filesystem code
