+++ StreamlineCore/src/slc/Common/Platform.h
@@ -59,3 +59,4 @@
     /* Unknown platform */
     #error "Unknown compiler!"
 #endif
+
