+++ TestApp/src/1BRC/BillionRows.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "BillionRowsV1.h"
+#include "BillionRowsV2.h"
+
+using BillionRows = BillionRowsV2;
