--- TestApp/src/BillionRows.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "BillionRowsV1.h"
-#include "BillionRowsV2.h"
-
-using BillionRows = BillionRowsV2;
